FAT-819:  Mark item withdrawn C10930
